requirements.txt
torch==2.0.1
torchvision==0.15.2
PyArabic==0.6.15
arabic-reshaper==3.0.0
six==1.16.0
numpy==1.23.5
pillow==10.2.0
tqdm==4.66.1
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
ultralytics==8.1.8
gradio==4.16.0
spaces==0.22.0
fastapi
uvicorn
